Fix export of `compareSpecificity`, fixes #323
